@@ -6,19 +6,22 @@
     "type": "project",
 
     "require": {
-        "php": ">=5.6.4",
-        "laravel/framework": "5.4.*",
-	"laravelcollective/html": "5.4.*",
-        "laravel/tinker": "~1.0",
-        "barryvdh/laravel-dompdf": "^0.8.0",
-        "intervention/image": "^2.3"
+	"php": ">=7.0",
+        "barryvdh/laravel-dompdf": "^0.8.2",
+        "intervention/image": "^2.4",
+        "laravel/framework": "5.5.*",
+        "laravel/tinker": "^1.0",
+        "laravelcollective/html": "5.5.*",
+        "spatie/flysystem-dropbox": "^1.0",
+        "spatie/laravel-backup": "^5.5"
     },
     "require-dev": {
-        "fzaninotto/faker": "~1.4",
+        "fzaninotto/faker": "~1.7",
         "mockery/mockery": "0.9.*",
-        "phpunit/phpunit": "~5.7",
+        "phpunit/phpunit": "~6.0",
         "symfony/css-selector": "3.1.*",
-        "symfony/dom-crawler": "3.1.*"
+        "symfony/dom-crawler": "3.1.*",
+	"filp/whoops": "~2.0"
     },
     "autoload": {
         "classmap": [
