@@ -25,12 +25,12 @@
         "classmap" : ["lib/"]
     },
     "require": {
-        "php": ">=5.3.0",
+        "php": ">=5.4.0",
         "ext-gd": "*",
         "ext-dom": "*",
         "ext-mbstring": "*",
         "phenx/php-font-lib": "0.5.*",
-        "phenx/php-svg-lib": "0.2.*"
+        "phenx/php-svg-lib": "0.3.*"
     },
     "require-dev": {
         "phpunit/phpunit": "4.8.*",
