@@ -15,26 +15,27 @@
         }
     ],
     "require": {
-        "php": ">=5.6.4",
+        "php": ">=7.0",
         "ext-mbstring": "*",
         "ext-openssl": "*",
-        "doctrine/inflector": "~1.0",
-        "erusev/parsedown": "~1.6",
-        "league/flysystem": "~1.0",
-        "monolog/monolog": "~1.11",
+        "doctrine/inflector": "~1.1",
+        "erusev/parsedown": "~1.7",
+        "league/flysystem": "^1.0.8",
+        "monolog/monolog": "~1.12",
         "mtdowling/cron-expression": "~1.0",
-        "nesbot/carbon": "~1.20",
-        "paragonie/random_compat": "~1.4|~2.0",
+        "nesbot/carbon": "^1.24.1",
+        "psr/container": "~1.0",
+        "psr/simple-cache": "^1.0",
         "ramsey/uuid": "~3.0",
-        "swiftmailer/swiftmailer": "~5.4",
-        "symfony/console": "~3.2",
-        "symfony/debug": "~3.2",
-        "symfony/finder": "~3.2",
-        "symfony/http-foundation": "~3.2",
-        "symfony/http-kernel": "~3.2",
-        "symfony/process": "~3.2",
-        "symfony/routing": "~3.2",
-        "symfony/var-dumper": "~3.2",
+        "swiftmailer/swiftmailer": "~6.0",
+        "symfony/console": "~3.3",
+        "symfony/debug": "~3.3",
+        "symfony/finder": "~3.3",
+        "symfony/http-foundation": "~3.3",
+        "symfony/http-kernel": "~3.3",
+        "symfony/process": "~3.3",
+        "symfony/routing": "~3.3",
+        "symfony/var-dumper": "~3.3",
         "tijsverkoyen/css-to-inline-styles": "~2.2",
         "vlucas/phpdotenv": "~2.2"
     },
