@@ -14,11 +14,12 @@
         }
     ],
     "require": {
-        "php": ">=5.6.4",
-        "illuminate/bus": "5.4.*",
-        "illuminate/contracts": "5.4.*",
-        "illuminate/queue": "5.4.*",
-        "illuminate/support": "5.4.*"
+        "php": ">=7.0",
+        "psr/log": "~1.0",
+        "illuminate/bus": "5.5.*",
+        "illuminate/contracts": "5.5.*",
+        "illuminate/queue": "5.5.*",
+        "illuminate/support": "5.5.*"
     },
     "autoload": {
         "psr-4": {
@@ -27,11 +28,11 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "5.4-dev"
+            "dev-master": "5.5-dev"
         }
     },
     "suggest": {
-        "pusher/pusher-php-server": "Required to use the Pusher broadcast driver (~2.0)."
+        "pusher/pusher-php-server": "Required to use the Pusher broadcast driver (~3.0)."
     },
     "config": {
         "sort-packages": true
