@@ -14,10 +14,10 @@
         }
     ],
     "require": {
-        "php": ">=5.6.4",
-        "illuminate/contracts": "5.4.*",
-        "illuminate/pipeline": "5.4.*",
-        "illuminate/support": "5.4.*"
+        "php": ">=7.0",
+        "illuminate/contracts": "5.5.*",
+        "illuminate/pipeline": "5.5.*",
+        "illuminate/support": "5.5.*"
     },
     "autoload": {
         "psr-4": {
@@ -26,7 +26,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "5.4-dev"
+            "dev-master": "5.5-dev"
         }
     },
     "config": {
