@@ -14,11 +14,10 @@
         }
     ],
     "require": {
-        "php": ">=5.6.4",
-        "illuminate/contracts": "5.4.*",
-        "illuminate/support": "5.4.*",
-        "nesbot/carbon": "~1.20",
-        "symfony/console": "~3.2"
+        "php": ">=7.0",
+        "illuminate/contracts": "5.5.*",
+        "illuminate/support": "5.5.*",
+        "symfony/console": "~3.3"
     },
     "autoload": {
         "psr-4": {
@@ -27,13 +26,13 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "5.4-dev"
+            "dev-master": "5.5-dev"
         }
     },
     "suggest": {
         "guzzlehttp/guzzle": "Required to use the ping methods on schedules (~6.0).",
         "mtdowling/cron-expression": "Required to use scheduling component (~1.0).",
-        "symfony/process": "Required to use scheduling component (~3.2)."
+        "symfony/process": "Required to use scheduling component (~3.3)."
     },
     "config": {
         "sort-packages": true
