@@ -38,7 +38,7 @@ return [
     |
     */
 
-    'debug' => env('APP_DEBUG', false),
+    'debug' => env('APP_DEBUG', true),
 
     /*
     |--------------------------------------------------------------------------
@@ -51,7 +51,7 @@ return [
     |
     */
 
-    'url' => env('APP_URL', 'http://localhost/aenaCIS'),
+    'url' => env('APP_URL', 'http://aenacis.aena.si'),
 
     /*
     |--------------------------------------------------------------------------
