@@ -1,68 +1,42 @@
-@extends('layouts.app')
+<!DOCTYPE html>
+<html lang="sl-SI">
+    <head>
+        <meta charset="utf-8">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+        {!! HTML::style('css/bootstrap4.css') !!}
+        {!! HTML::script('js/jquery.min.js') !!}
+        {!! HTML::script('js/bootstrap4.bundle.min.js') !!}
+        <title>{{ config('app.name', '') }}</title>
+    </head>
+    <body>
 
-@section('content')
-<div class="container">
-    <div class="row">
-        <div class="col-md-8 col-md-offset-2">
-            <div class="panel panel-default">
-                <div class="panel-heading">Login</div>
-                <div class="panel-body">
-                    <form class="form-horizontal" role="form" method="POST" action="{{ url('/login') }}">
-                        {{ csrf_field() }}
+        <div class="container">
 
+
+            <form class="form-horizontal" role="form" method="POST" action="{{ url('/login') }}">
+                {{ csrf_field() }}
+                <div class="card">
+                    <div class="card-header text-right">aena CIS</div>
+                    <div class="card-body">
+                        <h5 class="card-title text-center">PRIJAVA</h5>
+                        <div class="form-group text-right">
+                        </div>
                         <div class="form-group{{ $errors->has('email') ? ' has-error' : '' }}">
-                            <label for="email" class="col-md-4 control-label">E-Mail Address</label>
-
-                            <div class="col-md-6">
-                                <input id="email" type="email" class="form-control" name="email" value="{{ old('email') }}" required autofocus>
-
-                                @if ($errors->has('email'))
-                                    <span class="help-block">
-                                        <strong>{{ $errors->first('email') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
+                            <label for="email" class="control-label">E-Mail Naslov</label>
+                            <input id="email" type="email" class="form-control" name="email" value="{{ old('email') }}" required autofocus>
                         </div>
-
                         <div class="form-group{{ $errors->has('password') ? ' has-error' : '' }}">
-                            <label for="password" class="col-md-4 control-label">Password</label>
-
-                            <div class="col-md-6">
-                                <input id="password" type="password" class="form-control" name="password" required>
-
-                                @if ($errors->has('password'))
-                                    <span class="help-block">
-                                        <strong>{{ $errors->first('password') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
+                            <label for="password" class="control-label">Geslo</label>
+                            <input id="password" type="password" class="form-control" name="password" required>
                         </div>
-
-                        <div class="form-group">
-                            <div class="col-md-6 col-md-offset-4">
-                                <div class="checkbox">
-                                    <label>
-                                        <input type="checkbox" name="remember"> Remember Me
-                                    </label>
-                                </div>
-                            </div>
+                        <div class="form-group text-right">
+                            <button type="submit" class="btn btn-primary"> PRIJAVA </button>
                         </div>
-
-                        <div class="form-group">
-                            <div class="col-md-8 col-md-offset-4">
-                                <button type="submit" class="btn btn-primary">
-                                    Login
-                                </button>
-
-                                <a class="btn btn-link" href="{{ url('/password/reset') }}">
-                                    Forgot Your Password?
-                                </a>
-                            </div>
-                        </div>
-                    </form>
+                    </div>
+                    <div class="card-footer text-muted"> A1 Informatika d.o.o. / 064 170003 / info@aena.si</div>
                 </div>
-            </div>
+            </form>
         </div>
-    </div>
-</div>
-@endsection
+    </body>
+</html>
