@@ -3,7 +3,19 @@
 @section('content')
 
 
-<p><a class="btn btn-sm btn-primary" href="{!! URL::to('invoiceout/create') !!}">add new</a></p>
+{!! Form::open(array('route'=>'invoiceout.index','method' => 'get','class'=>'form-horizontal')) !!}
+<div class="row">
+    <div class="col-md-3  col-md-offset-6 ">{!! Form::select('idCustomer', $customer, $idCust, array('class' => 'form-control selChose')) !!} </div> 
+    <div class="col-md-2">{!! Form::select('year', $years, $year, array('class' => 'form-control')) !!} </div>
+    <div class="col-md-1">{!! Form::submit('Filter', array('class' => 'btn btn-success')) !!}</div>
+</div>
+{!! Form::close() !!}
+
+<div class="row">
+    <div class="col-sm-1 text-left"><a class="btn btn-sm btn-primary" href="{!! URL::to('invoiceout/create') !!}">dodaj račun</a></div>
+    <div class="col-sm-11 text-right"><?php echo $obj->appends(request()->input())->links(); ?></div>
+</div>
+
 
 <table class="table table-striped table-bordered">
     <thead>
@@ -11,7 +23,6 @@
             <td>ID</td>
             <td>Račun</td>
             <td>Datum oddaje</td>
-            <td>Rok plačila</td>
             <td>Podjetje</td>
             <td>Sklic</td>
             <td>Iznos NET</td>
@@ -24,14 +35,13 @@
             <td>{!! $value->id !!}</td>
             <td>{!! $value->nrInvoice !!}</td>
             <td>{!! $value->dateIssue !!}</td>
-            <td>{!! $value->dateDue !!}</td>
-          @if (isset($value->company->name))
-            <td>{!! $value->company->name  !!}</td>
+          @if (isset($value->cname))
+            <td>{!! $value->cname  !!}</td>
           @else
             <td> končni kupec </td>
           @endif
             <td>{!! $value->nrRef !!}</td>
-            <td>{!! $value->priceNet !!}</td>
+            <td>{!! DB::table('invoiceOutArt')->select(DB::raw('sum(priceUnit * qty) AS priceNet'))->where('idInvoiceOut',$value->id)->value('priceNet');!!}</td>
 
             <td class="text-right">
                 <a class="btn btn-xs btn-success" href="{!! URL::to('invoiceout/' . $value->id) !!}" target="_blank">S</a>
@@ -49,5 +59,8 @@
     window.onload = function() {
         sessionStorage.clear();
     }
+    
+    $('select.selChose').chosen({allow_single_deselect: true});
 </script>
+
 @stop
